Move ark related stuff out of main (#16)
